Allows kustomize build for all services
TL;DR
-----

Provides a kustomize file to include all services

Details
-------

Updates the `base` directory with a `kustomization.yaml` at that
level so a cluster can references that directory and install all
services. This is likely the way I'll plan on using the repository
for a long-running shared services cluster.